pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "signalation"
version = "0.0.0"
description = "Produce your messages from your signal account to kafka."
readme = "README.md"
authors = [{name="Markus Borea", email="borea17@protonmail.com"}]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.10"

dependencies = [
    "coloredlogs==15.0.1",
    "confluent-kafka==2.0.2",
    "click==8.1.3",
    "numpy==1.24.2",
    "pandas==1.5.3",
    "pydantic==1.10.5",
    "python-dotenv==1.0.0",
    "pyyaml==6.0",
    "requests==2.28.2",
]

[project.optional-dependencies]
dev = [
    "black==22.12.0",
    "flake8==6.0.0",
    "isort==5.11.4",
    "pre-commit==2.21.0",
    "pytest==7.1.0",
    "pytest-cov==3.0.0",
    "pytest-timeout==2.1.0",
    "pdbpp>=0.10.3",
    "requests-mock-1.10.0",
]

[project.urls]
Homepage = "https://github.com/borea17/signal-kafka-producer"

[project.scripts]
signal-kafka-producer = "signalation.services.producer:run"

[tool.setuptools]
include-package-data = true

[tool.setuptools.package-data]
"signalation.conf" = ["*.yaml"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = ['tests']
python_files = ['test_*.py']
